requirements.txt
streamlit
pandas
python-dotenv
requests
scikit-learn
plotly
wordcloud
matplotlib
Pillow
streamlit-autorefresh
